+++ src/DataConsumer.ts
@@ -87,12 +87,6 @@ export class DataConsumer extends EnhancedEventEmitter
 	// Closed flag.
 	private _closed = false;
 
-	// Buffered amount threshold.
-	private _bufferedAmountLowThreshold = 0
-
-	// Buffered amount.
-	private _bufferedAmount = 0
-
 	// Custom app data.
 	private readonly _appData?: any;
 
@@ -153,19 +147,6 @@ export class DataConsumer extends EnhancedEventEmitter
 	get dataProducerId(): string
 	{
 		return this._internal.dataProducerId;
-	}
-
-	/**
-	 * Buffered amount threshold.
-	 */
-	get bufferedAmountLowThreshold(): number
-	{
-		return this._bufferedAmountLowThreshold;
-	}
-
-	set bufferedAmountLowThreshold(value: number)
-	{
-		this._bufferedAmountLowThreshold = value;
 	}
 
 	/**
@@ -302,6 +283,64 @@ export class DataConsumer extends EnhancedEventEmitter
 	}
 
 	/**
+	 * Set buffered amount low threshold.
+	 */
+	async setBufferedAmountLowThreshold(threshold: number): Promise<void>
+	{
+		logger.debug('setBufferedAmountLowThreshold() [threshold:%s]', threshold);
+
+		const reqData = { threshold };
+
+		await this._channel.request(
+			'dataConsumer.setBufferedAmountLowThreshold', this._internal, reqData);
+	}
+
+	/**
+	 * Send data.
+	 */
+	async send(message: string | Buffer, ppid?: number): Promise<void>
+	{
+		if (typeof message !== 'string' && !Buffer.isBuffer(message))
+		{
+			throw new TypeError('message must be a string or a Buffer');
+		}
+
+		/*
+		 * +-------------------------------+----------+
+		 * | Value                         | SCTP     |
+		 * |                               | PPID     |
+		 * +-------------------------------+----------+
+		 * | WebRTC String                 | 51       |
+		 * | WebRTC Binary Partial         | 52       |
+		 * | (Deprecated)                  |          |
+		 * | WebRTC Binary                 | 53       |
+		 * | WebRTC String Partial         | 54       |
+		 * | (Deprecated)                  |          |
+		 * | WebRTC String Empty           | 56       |
+		 * | WebRTC Binary Empty           | 57       |
+		 * +-------------------------------+----------+
+		 */
+
+		if (typeof ppid !== 'number')
+		{
+			ppid = (typeof message === 'string')
+				? message.length > 0 ? 51 : 56
+				: message.length > 0 ? 53 : 57;
+		}
+
+		// Ensure we honor PPIDs.
+		if (ppid === 56)
+			message = ' ';
+		else if (ppid === 57)
+			message = Buffer.alloc(1);
+
+		const requestData = { ppid };
+
+		await this._payloadChannel.request(
+			'dataConsumer.send', this._internal, requestData, message);
+	}
+
+	/**
 	 * Get buffered amount size.
 	 */
 	async getBufferedAmount(): Promise<number>
@@ -346,20 +385,11 @@ export class DataConsumer extends EnhancedEventEmitter
 					break;
 				}
 
-				case 'bufferedamount':
+				case 'bufferedamountlow':
 				{
 					const { bufferedAmount } = data as { bufferedAmount: number };
-					const previousBufferedAmount = this._bufferedAmount;
 
-					this._bufferedAmount = bufferedAmount;
-
-					if (
-						previousBufferedAmount > this._bufferedAmountLowThreshold &&
-						this._bufferedAmount <= this._bufferedAmountLowThreshold
-					)
-					{
-						this.safeEmit('bufferedamountlow', bufferedAmount);
-					}
+					this.safeEmit('bufferedamountlow', bufferedAmount);
 
 					break;
 				}