+++ worker/src/RTC/DataConsumer.cpp
@@ -5,6 +5,7 @@
 #include "DepLibUV.hpp"
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
+#include "Utils.hpp"
 #include "Channel/Notifier.hpp"
 
 namespace RTC
@@ -90,6 +91,12 @@ namespace RTC
 
 		// Add protocol.
 		jsonObject["protocol"] = this->protocol;
+
+		// Add bufferedAmount.
+		jsonObject["bufferedAmount"] = this->bufferedAmount;
+
+		// Add bufferedAmountLowThreshold.
+		jsonObject["bufferedAmountLowThreshold"] = this->bufferedAmountLowThreshold;
 	}
 
 	void DataConsumer::FillJsonStats(json& jsonArray) const
@@ -146,6 +153,69 @@ namespace RTC
 				break;
 			}
 
+			case Channel::Request::MethodId::DATA_CONSUMER_SET_BUFFERED_AMOUNT_LOW_THRESHOLD:
+			{
+				auto jsonThresholdIt = request->data.find("threshold");
+
+				if (jsonThresholdIt == request->data.end() || !jsonThresholdIt->is_number_unsigned())
+					MS_THROW_TYPE_ERROR("wrong bufferedAmountThreshold (not an unsigned number)");
+
+				this->bufferedAmountLowThreshold = jsonThresholdIt->get<uint32_t>();
+
+				request->Accept();
+
+				break;
+			}
+
+			default:
+			{
+				MS_THROW_ERROR("unknown method '%s'", request->method.c_str());
+			}
+		}
+	}
+
+	void DataConsumer::HandleRequest(PayloadChannel::Request* request)
+	{
+		MS_TRACE();
+
+		switch (request->methodId)
+		{
+			case PayloadChannel::Request::MethodId::DATA_CONSUMER_SEND:
+			{
+				auto jsonPpidIt = request->data.find("ppid");
+
+				if (jsonPpidIt == request->data.end() || !Utils::Json::IsPositiveInteger(*jsonPpidIt))
+				{
+					MS_THROW_TYPE_ERROR("invalid ppid");
+				}
+
+				auto ppid       = jsonPpidIt->get<uint32_t>();
+				const auto* msg = request->payload;
+				auto len        = request->payloadLen;
+
+				if (len > this->maxMessageSize)
+				{
+					MS_WARN_TAG(
+					  message,
+					  "given message exceeds maxMessageSize value [maxMessageSize:%zu, len:%zu]",
+					  len,
+					  this->maxMessageSize);
+
+					return;
+				}
+
+				const auto* cb = new onQueuedCallback([&request](bool queued) {
+					if (queued)
+						request->Accept();
+					else
+						request->Error("message send failed");
+				});
+
+				SendMessage(ppid, msg, len, cb);
+
+				break;
+			}
+
 			default:
 			{
 				MS_THROW_ERROR("unknown method '%s'", request->method.c_str());
@@ -189,6 +259,30 @@ namespace RTC
 		MS_DEBUG_DEV("SctpAssociation closed [dataConsumerId:%s]", this->id.c_str());
 	}
 
+	void DataConsumer::SctpAssociationBufferedAmount(uint32_t bufferedAmount)
+	{
+		MS_TRACE();
+
+		auto previousBufferedAmount = this->bufferedAmount;
+
+		this->bufferedAmount = bufferedAmount;
+
+		// clang-format off
+		if (
+				previousBufferedAmount > this->bufferedAmountLowThreshold &&
+				this->bufferedAmount <= this->bufferedAmountLowThreshold
+		)
+		// clang-format on
+		{
+			// Notify the Node DataConsumer.
+			json data = json::object();
+
+			data["bufferedAmount"] = this->bufferedAmount;
+
+			Channel::Notifier::Emit(this->id, "bufferedamountlow", data);
+		}
+	}
+
 	// The caller (Router) is supposed to proceed with the deletion of this DataConsumer
 	// right after calling this method. Otherwise ugly things may happen.
 	void DataConsumer::DataProducerClosed()
@@ -204,7 +298,7 @@ namespace RTC
 		this->listener->OnDataConsumerDataProducerClosed(this);
 	}
 
-	void DataConsumer::SendMessage(uint32_t ppid, const uint8_t* msg, size_t len)
+	void DataConsumer::SendMessage(uint32_t ppid, const uint8_t* msg, size_t len, onQueuedCallback* cb)
 	{
 		MS_TRACE();
 
@@ -225,6 +319,6 @@ namespace RTC
 		this->messagesSent++;
 		this->bytesSent += len;
 
-		this->listener->OnDataConsumerSendMessage(this, ppid, msg, len);
+		this->listener->OnDataConsumerSendMessage(this, ppid, msg, len, cb);
 	}
 } // namespace RTC