+++ worker/src/RTC/PipeTransport.cpp
@@ -495,11 +495,11 @@ namespace RTC
 	}
 
 	void PipeTransport::SendMessage(
-	  RTC::DataConsumer* dataConsumer, uint32_t ppid, const uint8_t* msg, size_t len)
+	  RTC::DataConsumer* dataConsumer, uint32_t ppid, const uint8_t* msg, size_t len, onQueuedCallback* cb)
 	{
 		MS_TRACE();
 
-		this->sctpAssociation->SendSctpMessage(dataConsumer, ppid, msg, len);
+		this->sctpAssociation->SendSctpMessage(dataConsumer, ppid, msg, len, cb);
 	}
 
 	void PipeTransport::SendSctpData(const uint8_t* data, size_t len)